specific data_loading function in interface.
