+++ src/stimulus/cli/transform_csv.py
@@ -2,15 +2,14 @@
 """CLI module for transforming CSV data files."""
 
 import logging
-import os
 from typing import Any
 
 import datasets
 import pandas as pd
-import pyarrow as pa
 import yaml
 
 from stimulus.data.interface import data_config_parser
+from stimulus.data.interface.data_loading import load_dataset_from_path
 
 logger = logging.getLogger(__name__)
 
@@ -78,31 +77,6 @@ def transform_batch(
                 result_dict[column_name] = transform_obj.transform_all(result_dict[column_name])
 
     return result_dict
-
-
-def load_dataset_from_path(data_path: str) -> datasets.DatasetDict:
-    """Load dataset from various formats.
-
-    Args:
-        data_path: Path to the dataset (CSV, parquet, or HuggingFace dataset directory)
-
-    Returns:
-        A DatasetDict containing the loaded dataset
-    """
-    # Check if it's a directory (HuggingFace dataset)
-    if os.path.isdir(data_path):
-        logger.info(f"Loading dataset from directory: {data_path}")
-        return datasets.load_from_disk(data_path)
-
-    # Try to load as parquet first, then CSV
-    try:
-        logger.info(f"Attempting to load as parquet: {data_path}")
-        dataset = datasets.load_dataset("parquet", data_files=data_path)
-    except pa.ArrowInvalid:
-        logger.info("Data is not in parquet format, trying CSV")
-        dataset = datasets.load_dataset("csv", data_files=data_path)
-
-    return dataset
 
 
 def main(data_csv: str, config_yaml: str, out_path: str) -> None:
